存在チェックの方法変更

+++ src/main/kotlin/com/ss/poc/domain/order/IOrderCustomRepository.kt
@@ -3,7 +3,7 @@ package com.ss.poc.domain.order
 import com.ss.poc.application.order.ProductDto
 
 interface IOrderCustomRepository {
-    fun fetchCountOrderByOrderIdAndUserId(orderId: String, userId: Int, orderDetailId: String): Int
+    fun fetchCountOrderByOrderIdAndUserId(orderId: String, userId: Int, orderDetailId: String): Long
     fun updateOrderProductByOrderIdAndUserId(
         orderId: String,
         userId: Int,
